refactor: 인증 예외 클래스 및 패키지 구조 리팩토링
