--- src/main/java/roomescape/auth/application/exception/InvalidPasswordException.java
@@ -1,8 +1,0 @@
-package roomescape.auth.application.exception;
-
-public class InvalidPasswordException extends RuntimeException {
-
-    public InvalidPasswordException(String password) {
-        super();
-    }
-}
