+++ src/main/java/roomescape/auth/exception/InvalidEmailException.java
@@ -1,0 +1,8 @@
+package roomescape.auth.exception;
+
+public class InvalidEmailException extends AuthenticationException {
+
+    public InvalidEmailException(final String message) {
+        super(message);
+    }
+}
