+++ src/main/java/roomescape/auth/exception/InvalidPasswordException.java
@@ -1,0 +1,8 @@
+package roomescape.auth.exception;
+
+public class InvalidPasswordException extends RuntimeException {
+
+    public InvalidPasswordException(final String message) {
+        super(message);
+    }
+}
