+++ src/main/java/roomescape/auth/exception/TokenCreationException.java
@@ -1,0 +1,8 @@
+package roomescape.auth.exception;
+
+public class TokenCreationException extends RuntimeException {
+
+    public TokenCreationException(final String message) {
+        super(message);
+    }
+}
