+++ src/main/java/roomescape/auth/presentation/AuthController.java
@@ -6,16 +6,15 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
-import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.application.service.AuthService;
+import roomescape.auth.exception.UnauthorizedException;
 import roomescape.auth.presentation.dto.LoginCheckResponse;
 import roomescape.auth.presentation.dto.LoginRequest;
-import roomescape.global.exception.UnauthorizedException;
 
 @RestController
 @RequiredArgsConstructor
