+++ src/main/java/roomescape/global/exception/GlobalExceptionHandler.java
@@ -9,6 +9,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+import roomescape.auth.exception.TokenCreationException;
+import roomescape.auth.exception.UnauthorizedException;
 
 @RestControllerAdvice
 @Slf4j
